+++ NEWS.md
@@ -7,6 +7,7 @@ getopt 1.21.0 (development)
 * `getopt()` and `getusage()` now support action strings in column 3 of `spec`:
 
   + `"store"` stores argument value, legacy `1` supported as an alias.
+  + `"store_false"` stores `FALSE`.
   + `"store_optional"` stores argument value if present otherwise stores `TRUE`,  legacy `2` supported as an alias.
   + `"store_true"` stores `TRUE`, legacy `0` supported as an alias.